[BU000003ZA5DTC] Generate idempotency keys for POST requests outside of the retry loop
We recently added automatic generation of idempotency keys for POST requests and automatically retrying certain failed requests.

This was implemented slightly wrongly, as [noted](https://github.com/gocardless/gocardless-pro-python/issues/12#issuecomment-302194452) by teamup's @pegler, as we were generating the idempotency key inside the retry loop, meaning that we generated a new idempotency key for each request we retried, rather than using the same one. It's important that we use the same one as many of the failures we retry are network problems, where the request may have in fact made its way to the server and been successfully processed.

This extracts idempotency key generation to occur outside of the `for` loop for retries.

We also:

* refactor the configuration for retries is stored on the `BaseService`-inheriting object rather than passed around methods
* fix a bunch of tests that weren't getting run before (they had duplicate names, so only the last implementation was getting run)
* specify whether failures should be getting retried for a given API method in the service (some of the un-run tests we fixed were broken before, so we were accidentally retrying requests we didn't mean to be)